Add a new facility to "pin" the memory in an evbuffer chain.
For overlapped IO (and possibly other stuff) we need to be able to
label an evbuffer_chain as "pinned", meaning that every byte in it
must remain at the same address as it is now until it unpinned. This
differs from being "immutable": it is okay to add data to the end
of a pinned chain, so long as existing data is not moved.

git-svn-id: http://levent.svn.sourceforge.net/svnroot/levent/trunk/libevent@1142 ce677c24-8c1a-0410-9497-a30ef3a96221