Add a "ctrl" mechanism to bufferevents for property access.
OpenSSL uses something like this to implement get/set access for
properties on its BIOs, so that it doesn't need to add a pair of
get/set functions to the vtable struct for every new abstract property
it provides an accessor for.

Doing this lets us make bufferevent_setfd abstract, and implement an
abstract bufferevent_getfd.

git-svn-id: http://levent.svn.sourceforge.net/svnroot/levent/trunk/libevent@1284 ce677c24-8c1a-0410-9497-a30ef3a96221